chore: rename sample project for dealing with built in atlas textures
