+++ SampleProjects/Guides/BuiltInAtlasTextures/BuiltInAtlasTextures.csproj
