+++ SampleProjects/Guides/BuiltInAtlasTextures/Content/Atlas/atlas.png
