+++ SampleProjects/Guides/BuiltInAtlasTextures/stylecop.json
