+++ docs/guides/guides/rendering-atlas-textures.mdx
@@ -9,7 +9,7 @@ import StaticDownload from "@site/src/components/StaticDownload";
 # Rendering With An Atlas Texture
 
 :::note Guide Source Code
-Go to the <GuideLink projectName="AtlasTextures" /> guide project to see the source code for a fully working example of this guide.
+Go to the <GuideLink projectName="BuiltInAtlasTextures" /> guide project to see the source code for a fully working example of this guide.
 :::
 
 Another way to render textures is by using an atlas texture. This guide is a little more
